+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. See [conven
 ## Unreleased (main)
 #### Bug Fixes
 - leftover type-6 void HTML (`param`/`track`) does not swallow following prose
+- leftover wrap-created RST simple-table border skip-cut
 
 - - -